% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.all.schedule.R
\name{cpm.all.schedule}
\alias{cpm.all.schedule}
\alias{cpm_all_schedule}
\alias{genAllCpmSched}
\title{Generates all possible schedules for a cpm network}
\usage{
cpm.all.schedule(cpm)
}
\arguments{
\item{cpm}{CPM vector with Early Start Time, Early Finish Time,
Late Start Time and Late Finish Time.}
}
\value{
Matrix with all mininum makespan (the time to complete all jobs) cpm start time schedules
}
\description{
Generates all possible schedules for a cpm network
}
\examples{
#Use critical.path.method function to calculate a set of project
#activities:

ex.cpm.activities.duration <- c(1,4,5,7,2,3,1)
ex.cpm.activities.successors <- list(c(2,3), 4, c(4,5), 6, 7, 7, c(0))
ex.cpm <- cpm(ex.cpm.activities.duration,
              ex.cpm.activities.successors)

# Now, we have the CPM vector with:
# - est (Early Start Time)  - ex.cpm["est"]
# - eft (Early Finish Time) - ex.cpm["eft"]
# - lst (Late Start Time)   - ex.cpm["lst"]
# - lft (Late Finish Time)  - ex.cpm["lft"]

ex.cpm.activities.schedule <-
   cpm.all.schedule(ex.cpm)

}
\seealso{
Other scheduling: 
\code{\link{cpm}()},
\code{\link{mmf.all.sequences}()},
\code{\link{mmf.get.breakeven}()},
\code{\link{mmf.get.selffunding}()},
\code{\link{mmf.npv}()}
}
\concept{scheduling}
\keyword{activities}
\keyword{critical}
\keyword{path,}
\keyword{project}
\keyword{scheduling,}
