% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.discounted.cash.R
\name{draw.discounted.cash}
\alias{draw.discounted.cash}
\alias{draw_discounted_cash}
\title{Draw Discounted Cash vs Time}
\usage{
draw.discounted.cash(
  discounted.cash,
  smooth = 1,
  title = "Discounted Cash vs Time"
)
}
\arguments{
\item{discounted.cash}{A vector with discounted cash flow for each timestamp.}

\item{smooth}{Multiplier used to smooth line}

\item{title}{Used to define the title of plot}
}
\description{
Draw Discounted Cash vs Time
}
\seealso{
Other draw: 
\code{\link{draw.graph}()}
}
\concept{draw}
