% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.mmf.sf.be.R
\name{draw.mmf.sf.be}
\alias{draw.mmf.sf.be}
\alias{draw_mmf_sf_be}
\alias{drawMmfSfBe}
\title{TODO}
\usage{
draw.mmf.sf.be(df.complete)
}
\arguments{
\item{df.complete}{A compete dataframe.}
}
\description{
TODO
}
