% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.list.to.xls.R
\name{excel.list.to.xls}
\alias{excel.list.to.xls}
\alias{excel_list_to_xls}
\alias{excelListToXls}
\title{Export the generated ifm package results to a spreadsheet.}
\usage{
excel.list.to.xls(list.ifm.result)
}
\arguments{
\item{list.ifm.result}{The list with all results proccessed by the IFM
package.}
}
\value{
file.path The path to the generated file.
}
\description{
This function is responsible for reading a list of objects and
export a spreadsheet with the results proccessed by the IFM package.
The file contains:
\itemize{
 \item The raw data frame used to calculate the maxNPV, minSF and minBKE;
 \item The image oh the generated graph;
 \item The image oh the "Discounted Cash x Time" chart;
 \item The image oh the "MPV (ca$h) x Schedulling ID" chart;
 \item The image oh the "Self Funding (time) x Schedulling ID" chart;
 \item The image oh the "Breaking Event (time) x Schedulling ID" chart;
}
}
\examples{

ex.sheet.data <- excel.xls.to.list("../ifm/demo/resources/spreadsheet.xls")
ex.sheet.data.interest.rate <- ex.sheet.data[[1]]
ex.sheet.data.activities <- ex.sheet.data[[2]]
ex.sheet.data.durations <- ex.sheet.data[[3]]
ex.sheet.data.predecessors <- ex.sheet.data[[4]]
ex.sheet.data.cfs <- ex.sheet.data[[5]]

ex.mmf.seq <- mmf.all.sequences(ex.sheet.data.predecessors)

ex.mmf <- mmf.npv(ex.sheet.data.cfs,
                  ex.sheet.data.durations,
                  ex.mmf.seq,
                  ex.sheet.data.interest.rate)

ex.mmf.schedules <- ex.mmf[['schedules']]
ex.mmf.cfs.nominal <- ex.mmf[['cfs.nominal']]
ex.mmf.cfs.discounted <- ex.mmf[['cfs.discounted']]
ex.mmf.npv <- ex.mmf[['npv']]

ex.mmf.npv.selffunding <- mmf.get.selffunding(ex.mmf.cfs.discounted)
ex.mmf.npv.breakeven <- mmf.get.breakeven(ex.mmf.cfs.discounted)

ex.mmf.df.1r <- mmf.df.1r(ex.mmf.seq,
                          ex.mmf.schedules,
                          ex.mmf.npv,
                          ex.mmf.npv.selffunding,
                          ex.mmf.npv.breakeven)

ex.file.path <- excel.list.to.xls(ex.mmf.df.1r)

}
\seealso{
Other utility: 
\code{\link{excel.xls.to.list}()},
\code{\link{predecessors.to.edges}()},
\code{\link{utils.pred2graph}()},
\code{\link{utils.pred2suc}()},
\code{\link{utils.suc2pred}()}
}
\concept{utility}
\keyword{excel}
\keyword{list.to.xls}
\keyword{maxNPV}
