% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excel.xls.to.list.R
\name{excel.xls.to.list}
\alias{excel.xls.to.list}
\alias{excel_xls_to_list}
\alias{excelXlsToList}
\title{Import a list of variables from the spreadsheet to be used on the maxNPV
 function.}
\usage{
excel.xls.to.list(xls.spreadsheet.path)
}
\arguments{
\item{xls.spreadsheet.path}{The complete path to the spreadsheet that
represents the project.}
}
\value{
List of variables to be used on the maxNPV function.
}
\description{
This function is responsible for reading a spreadsheet
representing the project, and return a list with the following information
(in this order): The interest rate, the list of activities, the list of
durations of activities, the list of predecessors of activities, the
matrix that represents the cash flow series and the graph edges based on the
predecessor list
}
\examples{

print(getwd())
ex.sheet.data <- excel.xls.to.list("../ifm/demo/resources/spreadsheet.xls")


}
\seealso{
Other utility: 
\code{\link{predecessors.to.edges}()},
\code{\link{utils.pred2graph}()},
\code{\link{utils.pred2suc}()},
\code{\link{utils.suc2pred}()}
}
\concept{utility}
\keyword{excel}
\keyword{maxNPV}
\keyword{xls.to.list}
