% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation.free.interest.rate.R
\name{inflation.free.interest.rate}
\alias{inflation.free.interest.rate}
\alias{ifir}
\alias{IFIR}
\alias{IfIR}
\title{Calculate the Inflation-free Interest Rate.}
\usage{
inflation.free.interest.rate(interest.rate = 0.1425, inflation.rate = 0.0759)
}
\arguments{
\item{interest.rate}{A number that represents the nominal Interest Rate,
presented by year.}

\item{inflation.rate}{A number that represents the Inflation Rate, presented
by year.}
}
\value{
Returns the inflation-free interest rate
}
\description{
Calculate the Inflation-free Interest Rate.
}
\examples{

ex.ifir <- inflation.free.interest.rate(0.1425, 0.0759)

}
\seealso{
Other financial: 
\code{\link{discount.rate.vector}()},
\code{\link{discounted.csf}()},
\code{\link{draw.cfs}()},
\code{\link{net.future.value}()},
\code{\link{net.present.value}()}
}
\concept{financial}
\keyword{inflation-free,}
\keyword{interest}
\keyword{rate}
