% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmf.all.sequences.R
\name{mmf.all.sequences}
\alias{mmf.all.sequences}
\alias{mmf_all_sequences}
\title{Generates the topsorts list}
\usage{
mmf.all.sequences(predecessors = 0)
}
\arguments{
\item{predecessors}{List of Predecessors - Zero for none.  The index of the
list of predecessors represents the id of MMF and the value.}
}
\value{
List of all possible MMF sequences.
}
\description{
Generates the list of all possible MMF sequences (topsorts), 
constrained by the predecessors.
}
\examples{

ex.activities.predecessors<-list(0,1,2,3,1,5,c(4,6))
ex.mmf.seq <- mmf.all.sequences(ex.activities.predecessors)

}
\seealso{
Other scheduling: 
\code{\link{cpm.all.schedule}()},
\code{\link{cpm}()},
\code{\link{mmf.get.breakeven}()},
\code{\link{mmf.get.selffunding}()},
\code{\link{mmf.npv}()}
}
\concept{scheduling}
\keyword{activities}
\keyword{features,}
\keyword{marketable}
\keyword{minimum}
\keyword{project}
\keyword{scheduling,}
