% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmf.df.1r.R
\name{mmf.df.1r}
\alias{mmf.df.1r}
\title{Generates a data frame with Sequence, Schedule, NPV, Breakeven and
Self Funding}
\usage{
mmf.df.1r(mmf.seq, mmf.sched, mmf.npv, mmf.npv.selffunding, mmf.npv.breakeven)
}
\arguments{
\item{mmf.seq}{A list of sequences}

\item{mmf.sched}{A list of schedules}

\item{mmf.npv}{A list of NPV values}

\item{mmf.npv.selffunding}{A list of Selffunding times}

\item{mmf.npv.breakeven}{A list of Breakeven times}
}
\description{
Generates a data frame with Sequence, Schedule, NPV, Breakeven and
Self Funding
}
\examples{

data("ex.sheet.data")
ex.sheet.data.interest.rate <- ex.sheet.data[[1]]
ex.sheet.data.activities <- ex.sheet.data[[2]]
ex.sheet.data.durations <- ex.sheet.data[[3]]
ex.sheet.data.predecessors <- ex.sheet.data[[4]]
ex.sheet.data.cfs <- ex.sheet.data[[5]]

ex.mmf.seq <- mmf.all.sequences(ex.sheet.data.predecessors)

ex.mmf <- mmf.npv(ex.sheet.data.cfs,
                  ex.sheet.data.durations,
                  ex.mmf.seq,
                  ex.sheet.data.interest.rate)

ex.mmf.schedules <- ex.mmf[['schedules']]
ex.mmf.cfs.nominal <- ex.mmf[['cfs.nominal']]
ex.mmf.cfs.discounted <- ex.mmf[['cfs.discounted']]
ex.mmf.npv <- ex.mmf[['npv']]

ex.mmf.npv.selffunding <- mmf.get.selffunding(ex.mmf.cfs.discounted)
ex.mmf.npv.breakeven <- mmf.get.breakeven(ex.mmf.cfs.discounted)

ex.mmf.df.1r <- mmf.df.1r(ex.mmf.seq,
                          ex.mmf.schedules,
                          ex.mmf.npv,
                          ex.mmf.npv.selffunding,
                          ex.mmf.npv.breakeven)

}
