% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.future.value.R
\name{net.future.value}
\alias{net.future.value}
\alias{nfv}
\title{Net Future Value is a combination of different future values from different 
times, all which are put into one larger present value.}
\usage{
net.future.value(
  cfs = c(-350, 100, 200, 150, 75),
  interest.rate = 0.0619,
  begin.of.period = TRUE
)
}
\arguments{
\item{cfs}{A vector with a series of cash flows.}

\item{interest.rate}{A number that represents the nominal Interest Rate, 
presented by year.}

\item{begin.of.period}{A boolean that represents if the Tax Rate will be 
applied at the begining of period. FALSE by default.}
}
\value{
A future value of a cash flow series.
}
\description{
Net Future Value is a combination of different future values 
from different times, all which are put into one larger present value.
}
\examples{

ex.nfv <- net.future.value(c(-350,100,200,150,75), 0.0619, TRUE)

}
\seealso{
Other financial: 
\code{\link{discount.rate.vector}()},
\code{\link{discounted.csf}()},
\code{\link{draw.cfs}()},
\code{\link{inflation.free.interest.rate}()},
\code{\link{net.present.value}()}
}
\concept{financial}
\keyword{futureValue}
