% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net.present.value.R
\name{net.present.value}
\alias{net.present.value}
\alias{npv}
\title{Difference between the present values of cash inflows and outflows}
\usage{
net.present.value(
  cfs = c(-350, 100, 200, 150, 75),
  interest.rate = 0.0619,
  begin.of.period = TRUE
)
}
\arguments{
\item{cfs}{A vector with a series of cash flows.}

\item{interest.rate}{A number that represents the nominal Interest Rate,
presented by year.}

\item{begin.of.period}{A boolean that represents if the Tax Rate will be
applied at the begining of period. FALSE by default, the Tax Rate will be
applied to the second period.}
}
\value{
The sum of cash flows incomes/outcomes applying the Tax Rate to the
present time
}
\description{
calculates the difference between the present values of cash
inflows and outflows.
}
\examples{

ex.npv <- net.present.value(c(-350,100,200,150,75), 0.0619, TRUE)

}
\seealso{
Other financial: 
\code{\link{discount.rate.vector}()},
\code{\link{discounted.csf}()},
\code{\link{draw.cfs}()},
\code{\link{inflation.free.interest.rate}()},
\code{\link{net.future.value}()}
}
\concept{financial}
\keyword{presentValue}
