% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npvg.r
\name{npvg}
\alias{npvg}
\title{Net present value generalized.
Returns the net present value discount vector of a generalized cash flow stream.
input: relation csf={(vi,ti)} where vi is the element at time ti (byrow=F)
ti is an integer}
\usage{
npvg(csf, r = 0.01)
}
\arguments{
\item{csf}{cash flow stream vector}

\item{r}{rate}
}
\value{
generalized cash flow stream
}
\description{
Net present value generalized.
Returns the net present value discount vector of a generalized cash flow stream.
input: relation csf={(vi,ti)} where vi is the element at time ti (byrow=F)
ti is an integer
}
