% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predecessors.to.edges.R
\name{predecessors.to.edges}
\alias{predecessors.to.edges}
\alias{predecessors_to_edges}
\alias{predecessorsToEdges}
\title{Generate a vector with the edges to be plotted by draw.graph function.}
\usage{
predecessors.to.edges(list.of.predecessors = list())
}
\arguments{
\item{list.of.predecessors}{The vector that contain all activities predecessors.}
}
\value{
edges - List of edges to be used on the draw.graph function.
}
\description{
This function is responsible for reading a vector with all
activities predecessors and generate a list of the edges that will be
plotted by the draw.graph function.
}
\examples{

data("ex.sheet.data")
ex.edges <- predecessors.to.edges(ex.sheet.data$predecessors)

}
\seealso{
Other utility: 
\code{\link{excel.list.to.xls}()},
\code{\link{excel.xls.to.list}()},
\code{\link{utils.pred2graph}()},
\code{\link{utils.pred2suc}()},
\code{\link{utils.suc2pred}()}
}
\concept{utility}
\keyword{draw}
\keyword{draw.graph}
