% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schedules.1r.R
\name{schedules.1r}
\alias{schedules.1r}
\title{Generates all schedules for ONE resource, Denne Method.}
\usage{
schedules.1r(sequences, durations)
}
\arguments{
\item{sequences}{All sequences}

\item{durations}{Duration of activities}
}
\description{
Generates all schedules for ONE resource, Denne Method.
}
