% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.pred2graph}
\alias{utils.pred2graph}
\title{Plot activities.predecessors to graph.}
\usage{
utils.pred2graph(activities.predecessors)
}
\arguments{
\item{activities.predecessors}{The vector that contain all activities predecessors.}
}
\value{
Edges to plot a graph..
}
\description{
This function plot activities.predecessors to graph.
}
\examples{

ex.sheet.data <- excel.xls.to.list("../ifm/demo/resources/spreadsheet.xls")
g <- utils.pred2graph(ex.sheet.data$predecessors)
plot(g)

}
\seealso{
Other utility: 
\code{\link{excel.list.to.xls}()},
\code{\link{excel.xls.to.list}()},
\code{\link{predecessors.to.edges}()},
\code{\link{utils.pred2suc}()},
\code{\link{utils.suc2pred}()}
}
\concept{utility}
\keyword{activities}
\keyword{graph}
\keyword{predecessors}
\keyword{successors}
