% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.suc2pred}
\alias{utils.suc2pred}
\title{Transform activities.successors to activities.predecessors.}
\usage{
utils.suc2pred(activities.successors)
}
\arguments{
\item{activities.successors}{The vector that contain all activities successors.}
}
\value{
activities.predecessors.
}
\description{
This function convert activities.successors to
activities.predecessors.
}
\examples{

ex.sheet.data <- excel.xls.to.list("../ifm/demo/resources/spreadsheet.xls")
act.pred <- utils.suc2pred(ex.sheet.data$predecessors)

}
\seealso{
Other utility: 
\code{\link{excel.xls.to.list}()},
\code{\link{predecessors.to.edges}()},
\code{\link{utils.pred2graph}()},
\code{\link{utils.pred2suc}()}
}
\concept{utility}
\keyword{activities}
\keyword{predecessors}
\keyword{successors}
