% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{utils.pred2suc}
\alias{utils.pred2suc}
\title{Transform activities.predecessors to activities.successors.}
\usage{
utils.pred2suc(activities.predecessors)
}
\arguments{
\item{activities.predecessors}{The vector that contain all activities predecessors.}
}
\value{
activities.successors.
}
\description{
This function convert activities.predecessors to
activities.successors.
}
\examples{

ex.sheet.data <- excel.xls.to.list("../ifm/demo/resources/spreadsheet.xls")
act.suc <- utils.pred2suc(ex.sheet.data$predecessors)

}
\seealso{
Other utility: 
\code{\link{excel.xls.to.list}()},
\code{\link{predecessors.to.edges}()},
\code{\link{utils.pred2graph}()},
\code{\link{utils.suc2pred}()}
}
\concept{utility}
\keyword{activities}
\keyword{predecessors}
\keyword{successors}
